refactor: pipeline for scaffolding, ingest.mjs to retain existing ops/functionality
